Remove explict promise creation for ticket service

+++ server/services/ticket/TicketService.ts
@@ -1,8 +1,6 @@
 import '~/server/models/Notification';
-
 import Agenda from 'agenda';
 import { PubSubEngine } from 'apollo-server-express';
-import AWS from 'aws-sdk';
 import { UploadedFile } from 'express-fileupload';
 import S3 from '~/plugins/S3';
 import Analyst from '~/server/models/Analyst';
@@ -48,70 +46,55 @@ class TicketService {
     agenda.every('5 seconds', 'check sla');
   }
 
-  static getTickets(
+  static async getTickets(
     filter: any,
     sortBy: any,
     page: number,
     limit: number,
   ): Promise<Ticket[]> {
     // TODO sorting not works with doc ref
-    return new Promise((resolve, reject) => {
-      Ticket.find({}).then((tickets) => {
-        resolve(tickets);
-      });
-    });
+    return Ticket.find({});
   }
 
-  static getOne(ticketId: Ticket['id']): Promise<Ticket> {
-    return new Promise((resolve, reject) => {
-      Ticket.findOne(ticketId).then((ticket) => {
-        resolve(ticket);
-      });
-    });
+  static async getOne(ticketId: Ticket['id']): Promise<Ticket> {
+    const ticket = await Ticket.findOne(ticketId);
+    return ticket!;
   }
 
-  static copyTicket(ticketId: Ticket['id'], userId: Analyst['id']): Promise<Ticket> {
-    return new Promise((resolve, reject) => {
-      Ticket.findOne(ticketId).then(async (ticket) => {
-        const analyst = await Analyst.findOne(userId);
-        const newTicket = Ticket.create(ticket!);
-        newTicket.openedBy = analyst!;
-        newTicket.actualUser = analyst!;
-        resolve(TicketService.create(newTicket));
-      });
-    });
+  static async copyTicket(ticketId: Ticket['id'], userId: Analyst['id']): Promise<Ticket> {
+    const ticket = await Ticket.findOne(ticketId);
+    if (!ticket) throw new Error('Ticket not found');
+    const analyst = await Analyst.findOne(userId);
+    const newTicket = Ticket.create(ticket);
+    if (!analyst) throw new Error('Analyst not found');
+    newTicket.openedBy = analyst;
+    newTicket.actualUser = analyst;
+    return TicketService.create(newTicket);
   }
 
-  static updateOne(ticketId: Ticket['id'], ticketBody: Ticket): Promise<Ticket> {
-    return new Promise((resolve, reject) => {
-      Ticket.findOne(ticketId).then((ticket) => {
-        Object.assign(ticket, ticketBody);
-        ticket!.save().then(() => {
-          resolve(TicketService.getOne(ticketId));
-        });
-      });
-    });
+  static async updateOne(ticketId: Ticket['id'], ticketBody: Ticket): Promise<Ticket> {
+    const ticket = await Ticket.findOne(ticketId);
+    if (!ticket) throw new Error('Ticket not found');
+    Object.assign(ticket, ticketBody);
+    await ticket.save();
+    return TicketService.getOne(ticketId);
   }
 
-  static create(ticket: Ticket): Promise<Ticket> {
-    return new Promise((resolve, reject) => {
-      resolve(ticket.save());
-    });
+  static async create(ticket: Ticket): Promise<Ticket> {
+    return ticket.save();
   }
 
-  static changeStatus(
+  static async changeStatus(
     ticketId: Ticket['id'],
     statusId: Status['id'],
   ): Promise<Ticket> {
-    return new Promise((resolve, reject) => {
-      Ticket.findOne(ticketId).then(async (ticket) => {
-        const status = await Status.findOne(statusId);
-        ticket!.status = status!;
-        ticket!.save().then(() => {
-          resolve(ticket);
-        });
-      });
-    });
+    const ticket = await Ticket.findOne(ticketId);
+    const status = await Status.findOne(statusId);
+    if (!ticket) throw new Error('Ticket not found');
+    if (!status) throw new Error('Status not found');
+    ticket.status = status;
+    const savedTicket = await ticket.save();
+    return savedTicket;
   }
 
   static changeStatusOfTickets(
@@ -130,59 +113,51 @@ class TicketService {
     return Promise.all(promises);
   }
 
-  static transferToGroup(
+  static async transferToGroup(
     ticketId: Ticket['id'],
     groupId: Group['id'],
   ): Promise<Ticket> {
-    return new Promise((resolve, reject) => {
-      Ticket.findOne(ticketId).then(async (ticket) => {
-        const group = await Group.findOne(groupId);
-        ticket!.group = group!;
-        ticket!.actualUser = null;
-        ticket!.save().then((ticketSaved) => {
-          resolve(ticketSaved);
-        });
-      });
-    });
+    const ticket = await Ticket.findOne(ticketId);
+    const group = await Group.findOne(groupId);
+    if (!ticket) throw new Error('Ticket not found');
+    if (!group) throw new Error('Group not found');
+    ticket.group = group;
+    ticket.actualUser = null;
+    const ticketSaved = await ticket.save();
+    return ticketSaved;
   }
 
-  static commentOnTicket(
+  static async commentOnTicket(
     ticketId: Ticket['id'],
     userId: Analyst['id'],
     content: string,
   ): Promise<Comment> {
-    return new Promise((resolve, reject) => {
-      Ticket.findOne(ticketId, {
-        relations: ['comments'],
-      }).then(async (ticket) => {
-        const user = await Analyst.findOne(userId);
-        Comment.create({
-          user,
-          content,
-        })
-          .save()
-          .then((comment: Comment) => {
-            ticket!.comments.push(comment);
-            ticket!.save().then(() => {
-              resolve(comment);
-            });
-          });
-      });
+    const ticket = await Ticket.findOne(ticketId, {
+      relations: ['comments'],
     });
+    if (!ticket) throw new Error('Ticket not found');
+    const user = await Analyst.findOne(userId);
+    const comment = await Comment.create({
+      user,
+      content,
+    }).save();
+
+    ticket.comments.push(comment);
+    await ticket.save();
+    return comment;
   }
 
-  static removeChildren(
+  static async removeChildren(
     ticketId: Ticket['id'],
     children: Ticket['id'],
   ): Promise<Ticket> {
-    return new Promise((resolve, reject) => {
-      Ticket.findOne(ticketId, {
-        relations: ['children'],
-      }).then((ticket) => {
-        ticket!.children = ticket!.children.filter((child) => child.id !== children);
-        ticket!.save().then(() => resolve(TicketService.getOne(ticketId)));
-      });
+    const ticket = await Ticket.findOne(ticketId, {
+      relations: ['children'],
     });
+    if (!ticket) throw new Error('Ticket not found');
+    ticket.children = ticket.children.filter((child) => child.id !== children);
+    await ticket.save();
+    return TicketService.getOne(ticketId);
   }
 
   static async addChildren(ticketId: Ticket['id'], children: Ticket[]): Promise<Ticket> {
@@ -190,49 +165,48 @@ class TicketService {
     const ticket = await Ticket.findOne(ticketId, {
       relations: ['children'],
     });
-    ticket!.children.push(...children);
-    return ticket!.save();
+    if (!ticket) throw new Error('Ticket not found');
+    ticket.children.push(...children);
+    return ticket.save();
   }
 
-  static insertFile(ticketId: Ticket['id'], files: UploadedFile[]): Promise<File[]> {
-    return new Promise((resolve, reject) => {
-      Ticket.findOne(ticketId, {
-        relations: ['files'],
-      }).then(async (ticket) => {
-        files.forEach(async (f, index) => {
-          const name = `ticket/${ticketId}/${f.name} - ${index}`;
-          const params = {
-            Bucket: process.env.BUCKET!,
-            Key: name,
-            Body: f.data,
-          };
-          const { Location } = await S3.upload(params).promise();
-          const file = File.create();
-          file.name = name;
-          file.type = f.mimetype;
-          file.url = Location!;
-          await file.save();
-          ticket!.files.push(file);
-          await ticket!.save();
-          resolve(ticket!.files);
-        });
-      });
+  static async insertFile(ticketId: Ticket['id'], files: UploadedFile[]): Promise<File[]> {
+    const ticket = await Ticket.findOne(ticketId, {
+      relations: ['files'],
     });
+    if (!ticket) throw new Error('Ticket not found');
+    const result = await Promise.all(files.map(async (f, index) => {
+      const name = `ticket/${ticketId}/${f.name} - ${index}`;
+      const params = {
+        Bucket: process.env.BUCKET!,
+        Key: name,
+        Body: f.data,
+        ContentType: f.mimetype,
+        ACL: 'public-read',
+      };
+      const { Location } = await S3.upload(params).promise();
+      const file = File.create();
+      file.name = name;
+      file.type = f.mimetype;
+      file.url = Location;
+      return file.save();
+    }));
+    ticket.files = result;
+    await ticket.save();
+    return ticket.files;
   }
 
-  static getFile(fileId: string): Promise<Buffer> {
-    return new Promise((resolve, reject) => {
-      S3.getObject(
-        {
-          Bucket: process.env.BUCKET!,
-          Key: fileId,
-        },
-        (err: Error, file: AWS.S3.Types.GetObjectOutput) => {
-          if (err) reject(err);
-          return resolve(file.Body as Buffer);
-        },
-      );
-    });
+  static async getFile(fileId: File['id']): Promise<AWS.S3.Types.Body> {
+    const file = await File.findOne(fileId);
+    if (!file) throw new Error('File not found');
+    const { Body } = await S3.getObject(
+      {
+        Bucket: process.env.BUCKET!,
+        Key: file.name,
+      },
+    ).promise();
+    if (!Body) throw new Error('File not found');
+    return Body;
   }
 
   static async removeFile(
@@ -242,29 +216,28 @@ class TicketService {
     const ticket = await Ticket.findOne(ticketId, {
       relations: ['files'],
     });
+    const file = await File.findOne(fileId);
+    if (!ticket) throw new Error('Ticket not found');
+    if (!file) throw new Error('File not found');
     await S3.deleteObject({
       Bucket: process.env.BUCKET!,
-      Key: fileId.toString(),
+      Key: file.name,
     });
-    ticket!.files = ticket!.files.filter((f: any) => f.name !== fileId);
-    ticket!.save();
+    ticket.files = ticket.files.filter((f) => f.name !== file.name);
+    await ticket.save();
     return ticket!;
   }
 
-  static overtakeSla(ticketId: Ticket['id']): Promise<boolean> {
-    return new Promise((resolve, reject) => {
-      Ticket.findOne(ticketId).then((ticket) => {
-        resolve(ticket!.overtakeSla);
-      });
-    });
+  static async overtakeSla(ticketId: Ticket['id']): Promise<boolean> {
+    const ticket = await Ticket.findOne(ticketId);
+    if (!ticket) throw new Error('Ticket not found');
+    return ticket.overtakeSla;
   }
 
-  static slaPercentage(ticketId: Ticket['id']): Promise<Number> {
-    return new Promise((resolve, reject) => {
-      Ticket.findOne(ticketId).then((ticket) => {
-        resolve(ticket!.slaPercentage);
-      });
-    });
+  static async slaPercentage(ticketId: Ticket['id']): Promise<Number> {
+    const ticket = await Ticket.findOne(ticketId);
+    if (!ticket) throw new Error('Ticket not found');
+    return ticket.slaPercentage;
   }
 }
